requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
passlib[bcrypt]
bcrypt
fastapi-login
python-jose[cryptography]
python-dotenv
pydantic-settings
pytest
httpx
portalocker